+++ src/common/index.ts
@@ -1,10 +1,8 @@
 export * from "./ConductorLogger";
 export * from "./RequestCustomizer";
 export * from "./open-api/core/CancelablePromise";
-export * from "./orkes";
 export * from "./generators";
 export * from "./types";
-export * from "./sdk";
 
 export {
   ConductorClient,
