+++ src/core/sdk/dynamicFork.ts
@@ -1,4 +1,4 @@
-import { TaskType, ForkJoinDynamicDef, TaskDefTypes } from "../types";
+import { TaskType, ForkJoinDynamicDef, TaskDefTypes } from "../../common/types";
 
 export const dynamicForkTask = (
   taskReferenceName: string,
