+++ src/core/sdk/workflow.ts
@@ -1,4 +1,4 @@
-import { WorkflowDef, TaskDefTypes } from "../types";
+import { WorkflowDef, TaskDefTypes } from "../../common/types";
 
 export const workflow = (name: string, tasks: TaskDefTypes[]): WorkflowDef => ({
   name,
