+++ src/orkes/request/fetchCatchDns/index.ts
@@ -1,4 +1,3 @@
-export * from "./types";
 export * from "./FetchCache";
 export * from "./DnsCache";
 export * from "./fetchCatchDns";
